Moved changes from branch 77-address-issues-found-by-sara-code onto new develop branch

+++ lib/client/maat-client.h
@@ -105,7 +105,7 @@ int create_integrity_request(target_id_type_t target_typ,
                              xmlChar *cert_fingerprint,
                              xmlChar *info,
                              xmlChar **out,
-                             int *outsize);
+                             size_t *outsize);
 
 /**
  * Parse an integrity response contract. @input should be a serialized
@@ -136,7 +136,7 @@ int create_integrity_request(target_id_type_t target_typ,
  * TARGET_TYPE_UNKNOWN (for target_typ), and -1 is returned.
  */
 /* TODO Look at types of this function as compared to the ones in the create function */
-int parse_integrity_response(const char *input, int input_size,
+int parse_integrity_response(const char *input, size_t input_size,
                              target_id_type_t *target_typ,
                              xmlChar **target_id,
                              xmlChar **resource,
